Separate the command buffer and command encoder concepts
The CommandEncoder represents a top-level encoder of commands with a
finish() method that "consumes" the encoder and turns it into a command
buffer that is ready to be executed.

Advantages compared to the previous version of the IDL are:
 - Better type-safety: it shows the difference between the encoding of
commands and a finalized bundle of commands more explictly.
 - Better type-safety (for threads): only GPUCommandBuffers would be
transferrable / shareable to a different threads. Encoders would stay
inside the worker in which they were created.
 - More explicit validation point: some validation such as checking all
passes are ended, can only be done once we know the buffer is finished
encoding.
 - Allow implementations to free temporary encoding datastructures
early, which will be more important with reusable command buffers.

Disadvantages:
 - Requires an extra finish() call and introduces a mostly empty type,
increasing the complexity of the API slightly. (note that beginner WebGPU
developers can just do `queue.submit([encoder.finish()])` which doesn't
look very complicated)