Split AttachmentState between Color and DepthStencil states
The attachment state didn't contain much and the colorAttachments felt
redundant given the blendStates sequence had to have the same elements.

This renames GPUBlendStateDescriptor to GPUColorStateDescriptor and adds
the color attachment format to it. Likewise it adds the depth stencil
attachment format to the GPUDepthStencilStateDescriptor, and makes it
optional so that it can be null when there are no depth stencil
attahchments.