Add 'all' to GPUErrorFilter
It's pretty common for a programmer to want to capture all errors. Indeed, this is the behavior of glGetError().

Without this addition, an author wanting to do this would have to wrap their code in multiple error scopes,
one for each type of filter. This means that the WebGPU API can never add another type of error scope, because
existing code has to exhaustively list each one, and we can't expect every site to be updated when we add a new
type of error.