Remove createBufferMappedAsync
createBufferMappedAsync was added as a specific optimization for
multiprocess browsers with cross-process mapping (i.e. mmapping the
gpu process's buffer mapping into content-process memory).

This optimization is platform-dependent and difficult to implement. It
provides benefits for mapWriteAsync even without
createBufferMappedAsync. In absence of the optimization, there's no
reason to use createBufferMappedAsync, so it's unlikely that users will
do so.

So, to minimize the amount of unneeded API surface at 1.0, I think we
should take this out, and add it back once we have some implementation
experience, and more established WebGPU apps where we can test the
performance of a path like this.