Remove premerge
We've already forfeited perfect CFG round-trippability of SPIR-V because of the
lack of phi operations. So, now we're in a world of degrees, where we have to
decide whether a concept is important enough to be required to be exactly faithfully
preserved, or unimportant enough to allow the concept to be modified by the
round-trip.

The benefit of `premerge` is unclear, as it has yet to be demonstrated.
This concept doesn't exist in GLSL, HLSL, or MSL, so it is at least conceivable
that it is unnecessary.

On the other hand, it adds an unfamiliar keyword/block to the language,
which makes the language more difficult to read/write. Implementations will have to
implement it, test it thoroughly, figure out how to represent it in the platform
shading languages, and maintain it over time.

For `premerge` specifically, the current cost seems to outweigh the benefit, at least
until the benefit is demonstrated. This PR removes `premerge` until that happens.

closes https://github.com/gpuweb/gpuweb/issues/660