"zero"/null of storable types, rework initializers
Fixes #685

The generically zero-able types (or nullable types) correspond exactly with
storable types, so I tied nullability to the storable type concept.

Eliminate the concept of a "default" type.  Instead define the null
value of a storable type.

Clarify that a variable declaration can have an initializer.

Specify the initial value of a variable, and it's the null value for
the store type if the declaration had no intializer.

This still leaves up in the air exactly when a variable is "created"
as we haven't specfiied an execution order or even the concept of
execution of a statement.

Initializers are only permitted for 3 storage classes

Remove "Initializers" section later on, as its content is now mostly
abosrbed into the "Variables" section.

Hint at future reorganization, to insert a "Evaluation and Execution"
section between "Types" and "Grammar" major sections.