Small change to writeBuffer/Texture
This does the following changes to `writeBuffer/Texture`:
  - Reorder the arguments so that the destination comes first. Contrary
to copies that go from source to destination, this writes to an object
some data. So the object written to is more important.
  - Rename `source` to `data`.
  - Rename `destination` to `buffer` in writeBuffer.
  - `dataOffset` is made optional and defaulting to 0. It is also added
to `writeTexture` so that subparts of an `ArrayBuffer` can be selected
without having to create garbage. (and validation updated) (another
reason why the reordering was done)