Add GPULimits.maxUniformBufferBindingSize.
Start with a minimum maximum of 2^14 as that's the limit in Vulkan
that's hit for various ARM devices and for Swiftshader. Limits for Metal
and D3D12 are unclear but should be more than this.

The limit could be increased if we decided that the ARM devices should
be falling back to a WebGPU-compatibility feature level instead.