Remove is_finite and is_normal intrinsics.
This CL removes the is_finite and is_normal intrinsics. In order to
convert them into the OpIsFinite and OpIsNormal in SPIR-V the Kernel
capability is required. The Kernel capability requires a Kernel
execution mode. We do not use the Kernel execution mode but the
GLCompute execution mode for compute shaders.

If these methods are desired, they'll need to be added to the standard
library.