Make minBufferBindingSize optional
When the author opt-in to using minBufferBindingSize, this enables two checks:
1. The appropriate buffer in any bind groups created with this bind group layout is larger than minBufferBindingSize
2. minBufferBindingSize is larger than is necessary given the shader

If the author opts in to these requirements, we can do bounds checking at bind group creation / shader compilation time.

However, if the author does not opt in to these requirements, we must do bounds checking at draw call time.

Rather than using a sentinel 0 for the second check, we should have the author explicit enable them via out-of-band messaging.
This doesn't matter for the first check (since any buffer length is greater-than-or-equal-to 0), but it does matter for the
second check.

This patch doesn't change any defaults; it just makes makes the opt-in to earlier bounds checking more explicit.