Make BindGroupLayout parameters' default/undefined consistent
If the binding type in BindGroupLayout is incorrect, it is better
to say that hasDynamicOffset, minBufferBindingSize and multisampled
should be undefined. The current statement that an explicit default
value is correct doesn't make sense because the binding type is
totally wrong.

And this statement is consistent with other parameters like
viewDimension, storageTextureFormat.