Allow struct_decl to have a name.
This CL removes the requirement for a struct to be the RHS of a `type`
statement and, instead, allows an IDENT to be provided in the struct
declaration.

The struct is still required to be defined in the module scope.

The [[block]] attribute remains on the LHS of the struct_decl.

```
[[block]] struct A { [[offset 0]] b : f32; }
```

Fixes #735