GPUColor: remove sequence overload from the union
The overload adds complexity for not much benefit
(`{r: 0, g: 0, b: 0, a: 0}` vs `[0, 0, 0, 0]`).
More importantly, in certain contexts, the ordering of `[0, 0, 0, 0]` is
not obvious. For example, when used as the clear color of a BGRA
texture: does it use the "canonical" RGBA ordering, or does it match the
format? This removes the ambiguity.