Don't specify a strict grammar for decorators.
Decorations can't be validated much at parse time, so they shouldn't be
tightly specified in the grammar.

"[location(0), builtin(0)]" =parsing=> "Error: Failed to parse token after
builtin, expected IDENT."

"[location(0), builtin(frag_coord)]" =parsing=> OK =compilation=> "Error:
frag_coord cannot be an output of a vertex entrypoint."

Since having it in the grammar doesn't really reduce the validation load
on the compiler, we might as well not constrain it in the grammar. (even
if this double-validation is cheap/free)