[wgsl] Change SPIR-V memory model
This CL removes the requirement for the Vulkan memory model and
generates the `GLSL450` memory model. The Vulkan model isn't supported
over a large section of drivers yet and the original plan was to
generate and transform back to GLSL450 after the fact.

Instead, just generate straight to GLSL450 from WGSL.