Allow any value for bytesPerRow/rowsPerImage if they're unused
Semantically we really just want to allow bytesPerRow or rowsPerImage to
be omitted if it's unused. This strictly-loosens the validation so that
if unused, any number value is also allowed (instead of requiring that
it make sense for "the copy IF the {height,depth} were bigger than it
really is."

This should make the validation a little less annoying for users. E.g.
if they have corner cases in their stride computations for
{height,depth}==1, they won't break the copy for no good reason.

This is also motivated by simplifying the C API for webgpu.h users
(most importantly WASM applications running webgpu.h-on-WebGPU).
This lets them always specify `0`, so the C API doesn't have to be able
to express `undefined` as a separate value.