[wgsl] Add textureDimensions()
I've opted for 'Dimensions' over 'Size' as proposed in #1107. While more verbose, I think Dimensions is a better description as "size" may be mistaken for memory-size (especially with compute shaders, storage textures and overloads that return a single scalar).

I propose adding additional, separate functions for retrieving number-of-arrays, number-of-levels, number-of-samples, etc.

Bug: #1107