Change `offset` to `span`.
This CL removes the offset decoration from struct members and adds a
`span` decoration instead. The span reduces the amount of action at a
distance required when adding data members as each line is self
contained. This also reduces the amount of typing required as the span
is only required when the memory size does not patch the type size for a
member.

Fixes #1303