Rename GPUExtent3DDict.depth to .depthOrLayers
Updates Extent3D as well, as it would be confusing if they didn't match.

Keeps a placeholder "depth" member with type undefined, to prevent
accidental use of a silently-ignored dictionary entry.
Yes, this is weird, but WebIDL appears to allow it. If a browser
implementation of WebIDL doesn't support it - Chromium's doesn't - it
can be emulated by using another type (int, say) and then manually
throwing a TypeError if the member is present.