Replace `const` with `let`.
`let` is shorter and more tolerable to repetitively type than `const`.
`let` is used this way in Rust, which we already crib heavily from.
`let` does exist in JS, and although its use in WGSL will be more
similar to JS's `const`, we don't feel that this will be confusing.

We continue to reserve `const`.