Move clampDepth to GPUPrimitiveState.
GPUDepthStencilState contains state that's useful only when a depth
stencil attachment is present. However clampDepth clamps the depth
instead of clipping pixels even when there are no depth-stencil
attachments. That's why we move it to GPUPrimitiveState.

This is based on a comment @kvark made in passing that clampDepth should
be moved.