Add the definition of renderable format
It defines renderable format, and color renderable format, and
depth or stencil renderable format as well.

It removes one issue that rgb9e5unorm cannot be used as a color
attachment because rgb9e5unorm is not color renderable. Any
format has a -snorm suffix or any compressed format cannot used
as a color attachment because all of them are not color
renderable. We have already added this validation rule in
`GPURenderPassColorAttachment Valid Usage`. We don't need to
have separate text in the spec for rgb9e5unorm particularly.

renderable format definition is also needed for texture creation.
If a format is not a renderable format, it will fail with its
sampleCount > 1, or when its usage include `RENDER_ATTACHMENT`.
The later validation rule also make rgb9e5unorm + color attachment
combination fail.