Update GPUAdapterLimits to GPUSupportedLimits (#1765)
Previously the GPUDevice had a limits attribute that was of type object, but there doesn't seem to be any reason why it can't share an interface with the GPUAdapter's limits. To reflect that shared use that interface gets renamed to GPUSupportedLimits like the GPUSupportedFeatures before it.

Also added logic to clamp specified limits that were lower than the defaults to the better of the specified value and the default value, and made unrecognized limit keys a OperationError at Kai's suggestion.