Loosen pipeline fragment outputs validation (#1918)
This loosens validation so that fragment shaders do not need
to include an output location for color targets that cannot be written
to (have a writeMask of 0). This allows better shader reuse so that
the shader fragment outputs do not need to match the pipeline
color targets exactly if the pipeline does not intend to write
to those targets.