Make implictly created BGLs incompatible with all other pipelines. (#2057)
This change explicitly prevents GPUBindGroupLayouts created when no layout is specified during pipeline creation from being used with other pipelines. This is likely the most restrictive option that still keeps implicit BGL creation, but that means that it gives us a fair amount of leeway to loosen up restrictions in the future (maybe via a feature) without significant repercussions.