Add GPUComandEncoder.fillBuffer() (#2208)
Add GPUComandEncoder.fillBuffer() (#2208)

Fixes https://github.com/gpuweb/gpuweb/issues/1968.

In the future, we can add a 4th parameter, `value`, of type `octet` or `byte`. For now, though, the group seems to have just agreed on zeroing, rather than filling with a particular byte.

There are two commits in this PR, one that requires `GPUBufferUsage.STORAGE` on the buffer, and a second that relaxes this requirement because of https://github.com/gpuweb/gpuweb/issues/1968#issuecomment-886988850.