Define intrinsic size / natural dimensions of WebGPU canvases
This slightly refactors the state of GPUCanvasContext and makes a bunch
of small editorial changes in related areas.

Note the changes in GPUCanvasContext.configure which move all of the
error handling back to the content timeline (instead of the device
timeline). This is because it was previously attempted to early-return
from the device timeline and expected it to terminate the content
timeline steps, which isn't possible.