Move all canvas configure validation to getCurrentTexture
Previously, there was *some* validation inside configure(), but it only
covered a subset of the rules that were later covered by createTexture
inside getCurrentTexture(). This moves all of the validation to
getCurrentTexture() so it's not effectively duplicated.

Resolves an issue where configure() attempted to do device-timeline
validation but actually needed to do it on on the content timeline.

Extracted from #2416.