Add the concept of an "aspect-specific format" (#2702)
For example, if the app is populating a depth32float-stencil8 texture from a buffer, but is specifying
just the stencil aspect for the copy, then the data in the buffer should be packed uint8s (one for each
stencil value) without 4-byte holes between them. This is particularly important because it's legal to
copy to the stencil aspect of a depth24plus-stencil8 texture - so there can't be any holes because the
caller wouldn't know how big to make their holes.

This ends up interacting with validation. In [$validating linear texture data$], we make sure that
there's enough space for blockSize × widthInBlocks, so that blockSize needs to be 1 in the above case,
not 5. So, this PR modifies every call to [$validating linear texture data$] to pass the "aspect-
specific format" (which, in this example, turns the depth32float-stencil8 into just stencil8). That
way, the correct [=texel block size=] gets validated.

I believe this is almost entirely just editorial - there's one tiny non-editorial change in
writeTexture() that changes error handling. One of the prerequisites for calculating the aspect-specific
format is that you have to know the format is a valid image copy source or a valid image copy
destination, so this patch delays calling [$validating linear texture data$] until that validity
check has occurred.