Update logic for default TextureView dimension
Fixes #2684

Changed to consider the array layer count of the texture and
default to `2d-array` if the dimension is `2d` and the layer count
is greater than 1. Added a note to the `createView()` algorithm
pointing out that this may have unexpected results in data-driven
scenarios, so an specifying a dimension should be preferred.