[wgsl] Split regex apart by segment.
This PR displays the regex values by segment instead of as a continuous
string. The extract_grammar script is updated to rebuild the
non-separated regex before providing to tree-sitter.

The regex's are also exploded slight so there is more duplication
in them, but it means each segment stands alone to increase readability.