Set the default value of maxInterStageShaderComponents to 32
This is a stop-gap measure until https://github.com/gpuweb/gpuweb/issues/2749
is resolved. According to the Metal feature set tables, Macs limit the number
of inter-stage inputs to 32 and the number of inter-stage components to 124
If we limit the number of components to anything higher than 32, then a program
which uses single-component values will be able to exceed the limit on Macs.