[wgsl] Add some assert tags to the spec
This PR starts adding `assert` tags into the WGSL spec. The `assert` tag
is process by bikeshed which takes the text inside the `assert` and
generates a unique hash. The `assert` is then replaced by `span
id=assert-<hash>`. The hash will only change if the text inside the
`assert` tag changes.

This then allows tools to identify assertions in the text, and link to
them if desired.