Clarify type and use case for object labels.
Addresses part of #2633.

There were some concerns about the type (USVString or undefined) being
unusual for the web platform, so replaced it with simply USVString and
indicate that it defaults to empty string. Also added a note to explain
why USVString was used in favor of DOMString.

Additionally, there was some confusion about whether or not the label
was supplied by the browser or the developer and how it was used. Tried
to clarify the text to indicate that the label is developer-provided and
emphasize some possible use cases.