Make IO-shareable types apply to user-defined IO only.
The types of built-in pipeline IO values are fixed by the definition
of each particular built-in, so there's no need to have a category
that covers them. Restricting the definition of IO-shareable types to
exclude `bool` and `vec<bool>` makes the definition usable in the
description of user-defined IO.

Trade a Note about the special case of `bool` for a Note about
built-in types being independent of the IO-shareable category.