Move "expire stale external textures" slightly earlier
It was pointed out to me as written, a texture could be unexpired in any
event handlers called by the focus, resize, scroll, media query,
animation, fullscreen, and 2d context loss steps inside
"Update the rendering":
https://html.spec.whatwg.org/multipage/webappapis.html#update-the-rendering

This was put here because it lines up with requestVideoFrameCallback
(see the deleteed "Note:").
There's nothing fundamentally wrong with that, but this PR proposes
moving it earlier, so it runs _before_ "Update the rendering":

- It still runs after all user _tasks_.
- Avoids developers needing to understand the ordering of callbacks for
  different types of events (focus/resize/etc. vs rAF callbacks).
- No harm in it being earlier: the behavior is still tightly specified,
  and it lets us free the import earlier.

It's also perhaps a bit more future-proof: "Update the rendering" only
occurs in windows, not workers. It doesn't matter now, because only
HTMLVideoElement imports auto-expire. But if we added another
auto-expiring import in the future, this would be better.