Make GPUShaderModuleCompilationHint.layout optional
Issue #2574 shows that we may want to add more information to hints in
the future. To help make the hints future-proof, the layout is made
optional so that authors can in the future pass information like the
primitive type, even if they don't know what the pipeline layout will
be.