Throw when using formats without required feature
Adds algorithm steps to every method in the spec that accepts a
`GPUTextureFormat` causing them to throw a TypeError if the format
requires a feature which isn't enabled on the device. This is to
normalize behavior with implementations that don't implement the feature
at all, and will throw a type error at the IDL level.