Rename `discard` to `terminate`.
HLSL has "demote (to helper)" behavior for `discard`, but GLSL has
"terminate" behavior.
`terminate` is less ambiguous. "Terminate" is the verb used by e.g. SPIR-V, and it
more clearly contrasts with "demote" as a verb for the HLSL `discard`
behavior.

Also reserve `kill`, just in case.

This kind of reverts #918.