Fix and clean biblio references
Main fix prompted by recent updates in Bikeshed, which now fails to fix
inconsistent informative/normative references to a spec, see #2951.

Main fix is that explicit references to HTML were informative (using `[[HTML]]`)
instead of normative (using `[[!HTML]]`). This now seems to confuse Bikeshed,
and the HTML biblio entry appeared both under normative and informative
references as a result, with the same ID, which in turn made Bikeshed issue a
warning, preventing build (build is configured to fail on warnings)

The other updates are not needed to fix spec generation with Bikeshed but
further clean up biblio references:
- The term "resolve" in Web IDL can be referenced directly, no need to define it
locally (plus the spec URL was no longer current, as spec migrated to WHATWG)
- Canvas related terms needed to be associated with HTML no to create a [CANVAS]
entry in the index section that does not map to any [CANVAS] biblio entry.
- WGSL now exists in Specref, no need to have it in local biblio anymore (this
also allows to use the right "Working Draft" status for the spec, which is no
longer only an Editor's Draft)