Use PredefinedColorSpace instead of GPUPredefinedColorSpace
An issue indicates that doing this would mean that WebGPU automatically
gets support for new color spaces when additional entries are added to
PredefinedColorSpace. This seems like a good thing for the consistency
of the Web platform given WebGL and 2D canvas already do that. If a
future PredefinedColorSpace value is impossible to support in WebGPU
(which is unlikely), then additional validation can be added to the
spec.