Return null from getMappedRange if mapping is pending
getMappedRange can be used to detect whether a map has completed (even
before this change, using try/catch). This is a valid way of knowing
when a map is available - it's pretty much equivalent to
`mapAsync().then(() => { is_mapped = true; })`; the
state can only change when mapAsync's resolution happens anyway so the
state can't change inside a task.

This change makes this explicit by returning null instead of throwing an
exception in this case. This makes it distinguishable from real errors
which are still thrown as exceptions

Open questions:
- Should there be a way to query this without actually getting a mapped
  range? I think maybe, but you can always do the promise thing above.
- Should getMappedRange also return null when the buffer is in the
  unmapped state? I think no, it should still throw an exception.
- How does this get exposed in webgpu.h? (I'll open an issue there if we
  do this.)