Ensure that reflection attributes are initialized
Fixes #3064

This change ensures that even when object creation failes due to a validation
error the reflection attributes for that object are initialized to the values
from the descriptor for consistency. This happens to remove the need to
synchronize with the object creation on the device timeline before reporting
any of these values, which also seems like a good thing.