Stop using spec-prod for gh-pages builds
Using spec-prod for gh-pages builds means we push one commit to gh-pages
for each document being built in this repo. And each of these commits
@ mentions the person who clicked the merge button, sending them emails.

We already have a workflow for publishing docs to gh-pages because we
also want to publish webgpu.idl, and used to publish the specs that way.

I think the only thing this changes about our gh-pages output is that it
will have the "This version" line with the git hash in it like local
builds have now.