Remove styling for multiple <dd>s in a <dl>
The `display: list-item` was messing up numbering of surrounding `<ol>`
blocks (each `<dd>` item would increment the number of the outer list!)

Replace usages of multiple `<dd>`s in a row with a `<dd><ul>` (a list
inside a `<dd>`. Set the style of switches such that that doesn't result
in way too much indentation.