In certain places, identifiers don't resolve to declarations (#3117)
* Reorg declaration and scope

Reorder so we talk about all the kinds of declarations and their
scopes before talking about resolving an identifer.

Add the concept of, and a section for, context-dependent-name.
These are things that may look like an identifier, but only appear in
certain spots in the grammar, and in those spots never denote a declared
object. So the token doesn't undergo identifier/scope resolution.

Add a section to list context-depedent names in the language.

Created new grammar non-terminals so that 'ident' is reserved *solely*
for things that are the name for a declared object.

Fix for struct member names: They too don't use the normal
identifier/scope resolution.  So I added a member_ident grammar
nonterminal for this use.

Multi-letter swizzles are a single token

Fixes: #3113